Move `GeoShapeQueryRegionServiceTest` to `Region` folder
