+++ tests/ElasticSearch/Region/GeoShapeQueryRegionServiceTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\Search\ElasticSearch\Offer;
+namespace CultuurNet\UDB3\Search\ElasticSearch\Region;
 
 use CultuurNet\UDB3\Search\Region\RegionId;
 use Elasticsearch\Client;
